feat(Payment): Implement Stripe Checkout initiation Service

+++ DTOs/ReservationDTOs/ReservationDTO.cs
@@ -12,7 +12,7 @@ namespace GoDecola.API.DTOs.ReservationDTOs
         public DateTime CheckOutDate { get; set; }
         public DateTime ReservationDate { get; set; } = DateTime.UtcNow;
         public string? Status { get; set; }
-        public double TotalPrice { get; set; }
+        public long TotalPrice { get; set; }
         public UserDTO? User { get; set; } // propriedade de navegação para o usuário
         public TravelPackageDTO? TravelPackage { get; set; } // propriedade de navegação para o pacote de viagem
         public ICollection<GuestsDTO>? Guests { get; set; } // lista de hospedes associados à reserva
